requirements.txt
Django>=5.0,<6.0
django-import-export>=3.3,<4.0
mysqlclient>=2.2,<3.0
